requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
aiohttp
python-dotenv
httpx
markdown

# AI Libraries
openai
diffusers
torch
transformers
accelerate
soundfile